English Config Text Revising
Everything here is ok! All of my changes are on Keqing and Ningguang's loading info, mostly grammar corrections. There was also English that... I think it is technically correct, but sounds very unusual to a native English speaker, so I changed it a bit (mostly adding the word "are" in the case of Ningguang's loading info) so it sounds more natural.

I will do config text next, that will take some time so please wait for me to finish!
(I hope this is the correct way to make changes, I'm still familiarizing myself with GitHub)